+++ packages/frontend-v2/components/AddCourseModal/SelectedCourse.tsx
@@ -1,7 +1,7 @@
 import { Flex, Text } from "@chakra-ui/react";
 import { ScheduleCourse2 } from "@graduate/common";
 import { getCourseDisplayString } from "../../utils/";
-import { CourseTrashButton } from "../ScheduleCourse/CourseTashButton";
+import { CourseTrashButton } from "../ScheduleCourse/CourseTrashButton";
 
 interface SelectedCourseProps {
   selectedCourse: ScheduleCourse2<null>;
