+++ packages/frontend-v2/components/ScheduleCourse/CourseTrashButton.tsx
